# Optimize writing optional columns (part 2)