--- a/column_buffer.go
+++ b/column_buffer.go
@@ -92,6 +92,19 @@ func (a array) index(i int, size, offset uintptr) unsafe.Pointer {
 	return unsafe.Add(a.ptr, uintptr(i)*size+offset)
 }
 
+func (a array) slice(i, j int, size, offset uintptr) array {
+	if i < 0 || i > a.len || j < 0 || j > a.len {
+		panic("slice index out of bounds")
+	}
+	if i > j {
+		panic("negative slice length")
+	}
+	return array{
+		ptr: a.index(i, size, offset),
+		len: j - i,
+	}
+}
+
 type columnLevels struct {
 	repetitionDepth byte
 	repetitionLevel byte