Based on https://github.com/segmentio/parquet-go/pull/205, this PR optimizes the way we detect null values in rows.

I first attempted to optimize the `nullIndex*` and `nonNullIndex*` functions we were using but did not get a lot of success, then revisited the approach which yielded better results.

Instead of detecting contiguous ranges of null or non-null values, the algorithm now generates a bitmap indicating which values are nulls, and uses it to find the ranges. The main advantage is that generating the bitmap can be made a branch-less algorithm, which helps the CPU maximize throughput.

Once the bitmap is generated, the use of 64 bits words also help efficiently test ranges of null or non-null values; for example, with a simple comparison to zero we can determine whether a group of 64 values is all nulls.

The snapshot of benchmark below highlights the impact of the change on writing values to optional columns:
```
name                                      old time/op  new time/op  delta
GenericBuffer/optionalInt32Column/go1.18  1.98µs ± 0%  1.70µs ± 0%  -14.17%  (p=0.000 n=9+8)

name                                      old row/s    new row/s    delta
GenericBuffer/optionalInt32Column/go1.18    506M ± 0%    589M ± 0%  +16.50%  (p=0.000 n=9+8)
```